+++ project/src/main/java/syntactic/SyntacticMatcher.java
@@ -415,9 +415,9 @@ public class SyntacticMatcher
             return false;
     }
 
-    private float calculateWsScore(List<MatchedOperation> matchedOperations)
+    private double calculateWsScore(List<MatchedOperation> matchedOperations)
     {
-        float totalScore = 0;
+        double totalScore = 0;
 
         for (MatchedOperation mo : matchedOperations)
         {
@@ -427,9 +427,9 @@ public class SyntacticMatcher
         return totalScore / matchedOperations.size();
     }
 
-    private float calculateOpScore(List<MatchedElement> matchedElements)
+    private double calculateOpScore(List<MatchedElement> matchedElements)
     {
-        float totalScore = 0;
+        double totalScore = 0;
 
         for (MatchedElement me : matchedElements)
         {
